requirements.txt
absl-py==0.9.0
astunparse==1.6.3
atomicwrites==1.4.0
attrs==19.3.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
Cython==0.29.21
dill==0.3.2
future==0.18.2
gast==0.3.3
google-auth==1.20.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.31.0
h5py==2.10.0
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.0.1
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.3.1
more-itertools==8.4.0
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
Pillow==7.2.0
pluggy==0.13.1
promise==2.3
protobuf==3.12.4
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools-windows==2.0.0.2
pyparsing==2.4.7
pytest==6.0.1
pytest-mock==3.3.1
pytest-plt==1.1.0
python-dateutil==2.8.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
six==1.15.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow-datasets==3.2.1
tensorflow-gpu==2.3.0
tensorflow-gpu-estimator==2.3.0
tensorflow-metadata==0.23.0
termcolor==1.1.0
toml==0.10.1
tqdm==4.48.2
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
# pip install git+https://github.com/philferriere/cocoapi.git#subdirectory=PythonAPI